Change to use a Makefile to run the build.
While using a pure Python script is an interesting idea,
it does not handle detecting if dependencies are out of
date; since this is a core competency of make, it makes
sense to use make instead of Python to run the main
build process. If necessary, we can always write Python
scripts to handle any tricky parts.