Do not bundle python to a single file.
Bundle to a folder instead. The docs for pyinstaller say
that the single file command is slower, because when run all
the single file does is create a temp folder and then unpack
the bundled files into that folder before running. And we
are all about speed, so don't do that.