Make some things in ProgramInfo into Options.
The reasoning here is that even if we unwrap and use an empty string to
signify not found (which is lexically simpler) all we have done is introduced
a sentinel value that client code needs to check for. So really just the
same as using an Option<String> in the first place.